refactor(diffguard-core): split run_check into SRP submodules
The 165-line run_check function in check.rs did seven distinct things in
one body. Promote check.rs to a check/ module directory and extract each
concern into its own submodule:

- diff_prep      — parse, path-filter, allowed-line filter, dedup
- rule_filter    — tag-based rule selection (only/enable/disable)
- false_positive — fingerprint filtering + per-rule analytics tallies
- verdict        — status + exit code from counts
- annotations    — GitHub Actions annotation rendering
- rule_hits      — domain stats → core stats, with FP adjustments
- path_filter    — glob compilation + PathFilterError

run_check now reads as a 7-step pipeline. The public surface
(CheckPlan, CheckRun, RuleHitStat, PathFilterError, run_check) is
unchanged; sensor_api, the CLI, and the LSP keep working without
edits. Snapshot files relocated to check/snapshots/ to match insta's
source-relative discovery; their bodies are byte-identical.

Tests co-locate with the code they cover: tag-filter tests under
rule_filter, exit-code tests under verdict, the annotations proptest
under annotations. The run_check-level integration tests and snapshot
tests stay in check/mod.rs so snapshot file names remain stable.
